+++ src/shaders/glsl/collision_box.fragment.glsl
@@ -1,5 +1,5 @@
-flat in float v_placed;
-flat in float v_notUsed;
+in float v_placed;
+in float v_notUsed;
 
 void main() {