+++ src/shaders/glsl/line_gradient.fragment.glsl
@@ -1,6 +1,6 @@
 uniform sampler2D u_image;
 
-flat in vec2 v_width2;
+in vec2 v_width2;
 in vec2 v_normal;
 in float v_gamma_scale;
 in highp vec2 v_uv;